Disaplay account balance in nears in navigation
