+++ src/components/Navigation.js
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 
 import { NearContext } from '../contexts';
 
-import Dropdown from './common/Dropdown';
+import UserDropdown from './Navigation/UserDropdown';
 import Button from './common/Button';
 
 const Container = styled('div')`
@@ -53,34 +53,6 @@ const Container = styled('div')`
     color: var(--lavendar);
   }
 
-  .account-display {
-    color: var(--bubble-gum);
-    height: 52px;
-    line-height: 52px;
-  }
-
-  .account-display-id {
-    text-decoration: underline;
-  }
-
-  .dropdown-item {
-    :first-child {
-      margin-top: 17px;
-    }
-
-    :last-child {
-      margin-bottom: 17px;
-    }
-  }
-
-  .nav__link--dropdown {
-    display: block;
-    padding: 14px 30px;
-    color: var(--lavendar);
-    text-decoration: none;
-    white-space: nowrap;
-  }
-
   @media (min-width: 767px) {
     align-items: center;
     justify-content: space-between;
@@ -119,11 +91,7 @@ AccountDisplay.propTypes = {
 // );
 
 export default function Navigation() {
-  const { user, signOut } = useContext(NearContext);
-
-  const signOutAction = () => {
-    signOut();
-  };
+  const { user } = useContext(NearContext);
 
   const isHomePage = useRouteMatch('/').isExact;
   const isSignUpInPage = !!useRouteMatch('/sign-up') || !!useRouteMatch('/log-in');
@@ -143,18 +111,7 @@ export default function Navigation() {
         <div className="right">
           {/* eslint-disable-next-line no-nested-ternary */}
           {user ? (
-            <Dropdown dropdownBase={AccountDisplay} title={`${user.accountId}`} stretchable>
-              <Link className="nav__link nav__link--dropdown" to="/mint">
-                Mint a Gem
-              </Link>
-              <Link className="nav__link nav__link--dropdown" to="/profile">
-                View profile
-              </Link>
-              <span className="nav__link nav__link--dropdown">Balance: WIPⓃ ~$WIP USD</span>
-              <Link className="nav__link nav__link--dropdown" to="#" onClick={() => signOutAction()}>
-                Log out
-              </Link>
-            </Dropdown>
+            <UserDropdown />
           ) : (
             <>
               <Button isPrimary isLink>
